登录注册界面全变为了可滑动，首页定位城市，注册页面texyView填写时抬升，派单页面抬升仍有问题

+++ 鲁班同城/Category/AppDelegate+tabbar.m
@@ -162,7 +162,7 @@
            
         });
     } error:^(RCConnectErrorCode status) {
-         NSLog(@"登陆的错误码为:%d", status);
+        NSLog(@"登陆的错误码为:%ld", (long)status);
     } tokenIncorrect:^{
         //token过期或者不正确。
         //如果设置了token有效期并且token过期，请重新请求您的服务器获取新的token
